+++ packages/account-node/src/auto.ts
@@ -8,6 +8,7 @@ import VeridaDatabaseAuthType from "./authTypes/VeridaDatabase"
 import { AccountConfig, AccountNodeConfig, AuthContext, SecureContextConfig, SecureContextEndpointType, SecureContextServices, VdaDidEndpointResponses, VeridaDatabaseAuthTypeConfig } from '@verida/types'
 import { NodeSelector, NodeSelectorConfig, NodeSelectorParams } from './nodeSelector'
 import { ServiceEndpoint } from 'did-resolver'
+import { DefaultNetworkBlockchainAnchors } from '@verida/vda-common'
 
 /**
  * An Authenticator that automatically signs everything
@@ -26,11 +27,13 @@ export default class AutoAccount extends Account {
         super()
         this.accountConfig = accountConfig
         this.autoConfig = autoConfig
-        this.wallet = new Wallet(autoConfig.privateKey, <string> autoConfig.environment)
+
+        const blockchain = DefaultNetworkBlockchainAnchors[autoConfig.network]
+        this.wallet = new Wallet(autoConfig.privateKey, blockchain.toString())
 
         this.didClient = new DIDClient({
             ...autoConfig.didClientConfig,
-            network: autoConfig.environment
+            blockchain
         })
     }
 
@@ -77,7 +80,7 @@ export default class AutoAccount extends Account {
             },
             defaultNotificationServer:  {
                 type: 'VeridaNotification',
-                endpointUri: config.notificationEndpoints!
+                endpointUri: config.notificationEndpoints! ? config.notificationEndpoints! : []
             }
         }
     }
@@ -87,7 +90,7 @@ export default class AutoAccount extends Account {
             return this.defaultNodes
         }
 
-        config.network = this.autoConfig.environment
+        config.network = this.autoConfig.network
         config.defaultTimeout = config.defaultTimeout ? config.defaultTimeout : 5000
         config.notificationEndpoints = config.notificationEndpoints ? config.notificationEndpoints : []
 
@@ -102,7 +105,7 @@ export default class AutoAccount extends Account {
         await this.ensureAuthenticated()
 
         const did = await this.did()
-        let storageConfig = await StorageLink.getLink(this.didClient, did, contextName, true)
+        let storageConfig = await StorageLink.getLink(this.autoConfig.network, this.didClient, did, contextName, true)
         
         // Create the storage config if it doesn't exist and force create is specified
         if (!storageConfig && forceCreate) {
@@ -138,7 +141,7 @@ export default class AutoAccount extends Account {
      public async linkStorage(storageConfig: SecureContextConfig): Promise<boolean> {
         await this.ensureAuthenticated()
         const keyring = await this.keyring(storageConfig.id)
-        const result = await StorageLink.setLink(this.didClient, storageConfig, keyring, this.wallet.privateKey)
+        const result = await StorageLink.setLink(this.autoConfig.network, this.didClient, storageConfig, keyring, this.wallet.privateKey)
 
         for (let i in result) {
             const response = result[i]
@@ -157,7 +160,7 @@ export default class AutoAccount extends Account {
       */
     public async unlinkStorage(contextName: string): Promise<boolean> {
         await this.ensureAuthenticated()
-        let result = await StorageLink.unlink(this.didClient, contextName)
+        let result = await StorageLink.unlink(this.autoConfig.network, this.didClient, contextName)
         if (!result) {
             return false
         }
@@ -179,7 +182,7 @@ export default class AutoAccount extends Account {
      */
     public async linkStorageContextService(contextName: string, endpointType: SecureContextEndpointType, serverType: string, endpointUris: string[]): Promise<boolean> {
         await this.ensureAuthenticated()
-        const result = await StorageLink.setContextService(this.didClient, contextName, endpointType, serverType, endpointUris)
+        const result = await StorageLink.setContextService(this.autoConfig.network, this.didClient, contextName, endpointType, serverType, endpointUris)
 
         for (let i in result) {
             const response = result[i]