+++ packages/vda-name-client/src/blockchain/blockchainApi.ts
@@ -2,16 +2,16 @@ import {
     getVeridaContract,
     VeridaContract
 } from "@verida/web3"
-import { Web3SelfTransactionConfig, VdaClientConfig } from '@verida/types'
+import { Web3SelfTransactionConfig, VdaClientConfig, Network } from '@verida/types'
 import { ethers, Contract } from "ethers";
-import { getContractInfoForNetwork, RPC_URLS, getVeridaSignWithNonce } from "@verida/vda-common";
+import { getContractInfoForVeridaNetwork, RPC_URLS, getVeridaSignWithNonce, DefaultNetworkBlockchainAnchors } from "@verida/vda-common";
 import { JsonRpcProvider } from '@ethersproject/providers';
 import { explodeDID } from '@verida/helpers'
 
 export class VeridaNameClient {
 
     private config: VdaClientConfig
-    private network: string
+    private network: Network
     private didAddress?: string
 
     private vdaWeb3Client? : VeridaContract
@@ -38,7 +38,7 @@ export class VeridaNameClient {
             (<Web3SelfTransactionConfig> config.web3Options).rpcUrl = <string> RPC_URLS[this.network]
         }
 
-        const contractInfo = getContractInfoForNetwork("NameRegistry", this.network)
+        const contractInfo = getContractInfoForVeridaNetwork("NameRegistry", this.network)
 
         if (config.did) {
             this.readOnly = false
@@ -217,7 +217,8 @@ export class VeridaNameClient {
                 }
             }
 
-            const did = `did:vda:${this.network}:${response}`
+            const blockchain = DefaultNetworkBlockchainAnchors[this.network]
+            const did = `did:vda:${blockchain.toString()}:${response}`
             this.usernameCache[username] = did
             return did
         } catch (err:any ) {